reactor: expose cumulative tasks processed statistic
Count the total number of tasks the reactor processed and expose it
as a statistic. This is useful in controlled benchmarks to count
the number of tasks per unit of work (say a query), in order to check
for regressions or improvements.

The statistic is exposed in a new struct sched_stats, following
io_stats, to allow for expansion with other statistics in the
future.
Message-Id: <20210206211712.975107-1-avi@scylladb.com>