cli: Ignore the --num-io-queues option
The initial io groups patchset made this option configure the
number of io groups (with the deprecation warning). This was bad
idea, since this option often passes the total number of shards
on the system. In this case we get way too many io groups that
are not supposed to provide any reasonable qos.

Ignore the proposed value, but print a warning that the option
is deprecated and thus will be removed soon.

refs: #864

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>
Message-Id: <20210212090105.20751-1-xemul@scylladb.com>