rpc: streaming sink: order outgoing messages
The rpc streaming sink uses smp::submit_to() to send outgoing messages
to the shard the connection lives on to actually send them out. This
happens in the background the user of the sink is only blocked once the
memory limiting semaphore runs out of units. `smp::submit_to()` has no
ordering guarantee and indeed in debug build mode it was known to
reorder task, causing rpc to reorder messages. However recently we
observed this reordering in release mode as well. We don't know exactly
where and how it happens but it was tracked to the `smp::submit_to()`.
As there is opposition to give `smp::submit_to()` ordering guarantees,
the ordering guarantee has to be provided by the rpc sink, which is
exactly what this patch achieves.

Fixes #867.

Signed-off-by: Botond Dénes <bdenes@scylladb.com>
Message-Id: <20210215154601.288241-1-bdenes@scylladb.com>