file: expose disk offset alignment for overwrites
While read alignment can be a single sector, write alignment
is a single block. But starting with Linux 5.12, write alignment
on XFS is a single sector, provided the write is an overwrite -
a write to a location that was previously written, and not an
append.

This smaller alignment is useful for commit log files, as one
can collect a sub-block amount of data and write it to disk without
padding it to a full block size (only padding it to a single sector,
which is usually smaller).

To expose this, add a new file::disk_overwrite_dma_alignment()
method and propagate the values everywhere. Enable the relaxed
alignment for kernels >= 5.12 in query_dma_alignment().

A unit test is provided. I verified that it reports zero
stalls on a 5.10 kernel, many stalls with the alignment forced
to 512 on the same kernel, and 10 stalls for 5000 writes stalls on
a 5.12 kernel (with alignment reported as 512), which is an
acceptable result.
Message-Id: <20210223140239.4468-1-avi@scylladb.com>