rpc: print more info on bad protocol magic
In order to make it easier to figure out what is the source
of bad protocol magic in the RPC layer, the wrong magic is printed.
It can then be determined that if the bytes happen to look exactly
like first 8 bytes of an HTTP header, it's most likely the source
of confusion.
Example:
INFO  2021-02-23 15:18:01,731 [shard 0] rpc - client 127.0.0.1:40944: wrong protocol magic: 474554202f204854

Message-Id: <d80c25cb6262f99349ca77df66c2221164b8ee36.1614089916.git.sarna@scylladb.com>