reactor: reactor_backend_epoll: stop using signals for high resolution timers
The epoll backend uses signals for the high resolution timer,
since we want the high resolution timer to preempt running tasks,
and not just be recognized on a task-quota boundary. This was first
just a nice-to-have, then became important in 6c97dcf4bda52
("thread: scheduling groups") where we allowed specifying a quota
for a specific seastar::thread using a thread_scheduling_group;
this required sub-task-quota scheduling. It became unimportant
again in c8eb38d7e449 ("thread: remove deprecated
thread_scheduling_group and its unit test"), as there was no longer
any critical user of tsub-task-quota scheduling. However, it has
now become important again with 96562959be3 ("ioq: Prepare the ioq
poller for wakeup"), when I/O scheduling needs sub-task-quota
precision.

While the current signal-based high resolution timer system works,
it uses the non-scalable signal subsystem in Linux, which synchronizes
using a process-wide lock. We've previously seen terrible behavior
on large systems with cross-reactor wakeup notifications (see
1c2917e9bd71 ("reactor: switch smp notifications from signals to
eventfds") and c175cb527806 ("Merge "reactor: stop using signals
for task_quota timer"")). Although I/O scheduling is less likely to
have wakeups at the same time, and thus trigger contention, it's better
to stop using signals for timers.

The solution in this patch is to have the timer thread (used for
task-quota preemption) also be responsible for high resolution
timer wakewups, by using a timerfd instead of a posix timer and
using poll() to wait for either a task quota wakeup or a high
resolution timer wakeup.

One complication is that when the reactor sleeps, we cannot
use the timerfd as a wakeup source, as the timer thread could
pick up the wakeup first, and we'd miss a wakeup. The somewhat
ugly solution is to use two timerfds, having only one enabled
at any time. This causes 4 extra syscalls any time we go to sleep,
but that is a slow path.

Note the whole thing only matters for the epoll backend, which
is used on pre 4.18 kernels. Newer kernels use the linux-aio
backend, which is able to use a single timerfd for the task.

Test: unit(dev)
Message-Id: <20210227124600.1476555-2-avi@scylladb.com>