log: skip on unknown nested mixing instead of stopping the logging
The logging function for an exception type will stop if a nested
exception change is not mixed in with a type derived from
std::exception. The consequence of this is that if there are more
nested exceptions in the chain they will not be printed out which
means that we can lose viable debugging information.
This commit fixes this by adding a catch handler that will skip the
unknown part of the chain and will try to continue to the next part of
the exception chain which is the std::nested_exception member.

It also adds explicit indication in the logging where there is no
information on an object - for example when it is not derived from
std::exception.

Tests for the exception logging are added for all of the different
cases.

Fixes #874