utils/log.cc: fix nested_exception logging (again)
After multiple attempts nested_exception logging is still not robust.

In particular if std::system_error appears in the middle on the nested
exception chain the logging will break and very important information
(e.g. the object that initiated the exception chain) is going to be lost.

This patch is going to syntactically split the "printing part" of the
operator<<(exception_ptr) and the recursion part where we unroll
the std::nested_exception object.

As a result it's going to become much harder to break this code in the
future.

Signed-off-by: Vlad Zolotarov <vladz@scylladb.com>