file, reactor: reinstate RWF_NOWAIT support
RWF_NOWAIT support was disabled in e67572ef68a6cd1f0 ("reactor:
disable nowait aio due to a kernel bug"). It was later re-enabled
but disabled again.

The kernel problems have now been indentified[1] and so we can
can re-enable it. Since each filesystem was fixed in a different
kernel version, we need to plumb down this information from
posix_file_impl and friends all the way down to the reactor backend,
via io_request. This is annoying but cannot be helped.

Tested via the unit tests and also verified that the nowait paths
are exercised with strace.

[1] https://lore.kernel.org/linux-xfs/20210117213401.GB78941@dread.disaster.area/
Message-Id: <20210315085935.1260903-1-avi@scylladb.com>