reactor: add --kernel-page-cache option to disable O_DIRECT
O_DIRECT is great when an application manages a long-lived cache
and doesn't want the kernel page cache to interfere. It's not so
great with short-lived unit tests, since they usually don't have
caches, and when they do, they aren't populated.

Enabling the page cache allows unit test to read input data from
memory instead of from disk, and writes can also stay in memory
provided the file is deleted quickly (and provided
--unsafe-bypass-fsync 1 is also supplied).

Linux-aio is not asynchronous when the page cache is enabled, so
we use the syscall thread (luckily, we can reuse the code for
EAGAIN when RWF_NOWAIT is used). This isn't super performant
since requests will issue serially, but we don't expect high
concurrency for unit tests and we do expect most reads and writes
to hit cache.

Test: unit (dev), strace for file_io_test.
Message-Id: <20210317201131.219783-1-avi@scylladb.com>