file: Add option to refuse the append-challenged file
By default all files that are opened for write (or read-write)
on a filesystem that's considered to handle appends poorly (XFS)
are forced to use the append-challenged file implementation.
Note, that the challenged implementation doesn't come for free,
it uses notable more cpu cycles than plain posix file one and
may need to allocate memory.

In scylla, however, commitlog truncates the file to its desired
size or fills it with zeroes by hands right at once, and then
writes within the truncated bounds. Trailing zeroing, if happens,
goes with a single request which also doesn't need any special
handling even on XFS. In other words commitlog doesn't need the
append challenge to be solved, but it cannot refuse the cpu/mem
overhead that's imposed.

The proposal is to let caller hint open_file_dma that it's not
going to do write-behinds and thus the append optimization and
the associated cost is not required.

tests: scylla.unit(dev)

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>
Message-Id: <20210318094043.7946-1-xemul@scylladb.com>