iotune: Estimate accuarcy of measurement
The iotune currently runs 4 measurements. For each it submits
a bunch of requests and reports back the simple average of the
measured value as the "capacity" (whatever it means). While we
don't have anything better, it's worth to know the accuracy of
such a measurement.

Respectively, this patch teaches iotune to collect the number
of requests each measurement submits per second, calculate the
standard deviation of the collected arrays and report if the
deviation exceeds a certain threshold (defatul is 3%). Large
deviation is not treated as error, but is a sign to the caller
that disk may behave not the way io-scheduler expectes it to
and needs more attention.

This trick would have helped to note the instability of google
disks -- currently it shows 5% (and more) deviation of write
throughput measurement for the write-back cache and 0.5% for the
write-through one.

On AWS disks it surprisngly shows pretty unstable IOPS with ~10%
deviation. This is to be looked at later.

For other parameters the observed deviation is less than 0.5%.

tests: iotune on aws, gce and local disk

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>
Message-Id: <20210319082719.26766-1-xemul@scylladb.com>