io_queue: Provide request limits
The queue works with any requests, but if a request length exceeds
a certain threshold a warning is emitted and the request ticket is
trimmed down. The latter is not fatal, but results in increased IO
latencies.

To let caller reason about IO lengths report this threshold back.

The IO queue has internal math for calculating tickets out of
request lenghts. This implementation detail is kept private, and
the lenghts for read and write requests are reported.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>