cmake: filter out -Wno-error=#warnings from pkgconfig (seastar.pc)
The clang -Wno-error=#warnings flag, used to not break the build when
a #warning directive is encountered, is unfortunate in containing
a comment introducing character ("#"). In our case, pkg-config
interprets it as a comment and truncates the line containing it at
that position.

Unfortunately, we cannot bring to bear cmake's text processing functions
(for example, string (REPLACE ...)) since compile options are not
available during configure time (see [1]). We can process compile options
using generator-expressions (see [2]), but these facilities do not
include a general search-and-replace.

To allow forward process, simply filter out the annoying option.
This is doable using generator-expressions. The pkgconfig output
is now usable, but users will have to detect the availability of
-Wno-error=#warnings and add it themselves and cannot rely on Seastar
to provide it. This only matters on platforms where Seastar issues
such warnings (anything that isn't x86_64, aarch64, s390x).

[1] https://stackoverflow.com/questions/51353110/how-do-i-output-the-result-of-a-generator-expression-in-cmake
[2] https://cmake.org/cmake/help/latest/manual/cmake-generator-expressions.7.html#string-transformations
Message-Id: <20210321130131.1801926-1-avi@scylladb.com>