file: Add RFW_NOWAIT detection case for AuFS
Fixes #883

Found when trying to run scylla inside a container in a not-so-new docker, using
AuFS as container layer fs. This setup does not support RWF_NOWAIT, regardless (?)
of kernel version. At the very least not at 4.15.

This is a pessimistic fix, if there is a version or combination thereof where it
works, please add to this. But since even docker is phasing out AuFS, it might
not be very important. Except for me and my old container.

Message-Id: <20210329133204.19493-1-calle@scylladb.com>