tls: preserve exact error state so repeated calls generate same message
Fixes #885

In some error cases, we set the state of a TLS session to "error".
Any further attempts to use the session will result in failure.
We should preserve the exact reason for this error and report
it on repeated queries, instead of just "broken".

Fixed by making "bool _error" into a std::exception_ptr instead.
This can be set by either gnutls error codes or underlying
transport exceptions.
Message-Id: <20210331133331.16598-1-calle@scylladb.com>