io_tester: Add option to reuse test files
When io-tester is run several times in a row it makes sense
not to remove and create files from scratch. This is a huuuge
and absolitely safe time saving for large files and read
workloads. Write ones should use it with care, but given long
enough testing time it's also safe.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>