test: rpc_test: reduce memory usage in compression tests
rpc_test fails on small memory setups (or setups with
~/.config/seastar/seastar.conf restricting memory usage). The problem is
the compression tests construct large test vectors that exhaust memory.

Fix by changing the test vector to hold a function that generates the
test input, rather than hold the input itself. This means we need to
hold the input for just one test in memory at any one time.

Even with this, memory exceeds 1GB when we split an 8MB input into
1-byte temporary buffers. I dropped the 8MB tests, I don't think testing
8MB inputs adds confidence that 4MB inputs don't bring.
Message-Id: <20210405120730.271370-1-avi@scylladb.com>