dns: fix tcp sendv return value to c-ares
We return `len`, which is the length of the iovec, instead of the
number of bytes sent. This causes c-ares to think it sent just
one byte, so it sends the entire packet (minus the first byte)
again, and then again.

UDP works on packets, so it ignores partial writes. This allows
UDP named resolution to work.

Fix by returning the number of bytes in the packet. Note we don't
have the ability to report a partial send, so we report either
EWOULDBLOCK or that we buffered the entire sendv.

I added a return for the case where buffer space was available,
even though fall-through would work. I think it's clearer as
we terminate that path early rather than go through a meaningless
check for f.failed().

The unit test is reinstated. Note another tcp test somehow
worked despite the bug on my machine.

Fixes #521 (again).
Message-Id: <20210409074948.391338-1-avi@scylladb.com>