io_queue: Do not use write bandwidth/iops as limit
To limit requests scheduler operates on two values:
- limit
- write-multiplier (just multiplier below)

The limit for both -- throughput and iops -- is currently taken
as the maximum of read and write measurements. The multiplier is
calculated as read_speed/write_speed and is treated as "weight"
of a write request, while reads a weighed as 1. When scheduling
the sum of weights is maintained not to go above the limit.

This math works as expected when read is faster than write. E.g.
let's take a disk with read iops 200 and write iops 100. The limit
is then  200 and the multiplier is 2. Respectively, the scheduler
allows for 200 / 1 = 200 reads and 200 / 2 = 100 writes.

This math falls apart when the disk write speed is larger than its
read speed (not sure if they exist though). Reversing the above
example, if a disk is 100 read iops and 200 write iops the limit
would be 200 again, but the multiplier would become 0.5. Next, the
number of reads is 200 / 1 = 200 and the maximum number of writes
is 200 / 0.5 = 400. Disk is overloaded 2 times.

The fix is to treat read speed as the limit and use write speed
only to get the multiplier.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>