iostream: Use options struct construct output_stream
The output_stream has 1 mandatory parameter -- the buffer size -- and
2 optional, both are booleans. This patch adds an options strucuct
to initialize the latter. It serves two purposes -- make options names
explicit and helps to avoid combinatorial explosion of options (in the
future, if any). The buffer size is not on the options because it's
not an option and doesn't have any default value.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>