iotune: Measure disk saturation lengths
To measure the saturation length the test starts with the maximum
request size (1M) and halves it several times until the measured
thrgouthput falls below the original value's deviation.

It's critical to measure saturation before the disk is randomized
with IOPS test, so write piece comes right after the write bandwidth
test, the read piece after the read bandwidth one.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>