io_queue: Count saturation length in max request lenghts
When chosing the maximum request length one needs to consider
two corner cases.

Limiting request lenghts with smaller values is better from the
latency perspective. Smaller requests complete sooner making the
scheduler life simpler. Also, and more important, smaller requests
are more mixed-workload friendly -- we've seen reads and writes
of 128k both may have no chances to meet the latency goal. On the
other hand mixing 64k requests makes it possible (not yet coded).

Limiting requests lengths with larger values is better from the
throughput perspective. E.g. on modern cloud NVME SSDs it's not
possible to saturate the disk bandwidth with requests smaller than
16k. Also, the larger requests are the less of them is needed to
write a certain amount of data, the less CPU ticks it takes.

The decision that should suit both parties is to limit requests
lengths with the minimal saturation lengths. We'll still be able
to squeeze as much as we can from the disk bandwidth and also
current scheduler will serve mixed workload with better latencies.

Later, when the proper handling of mixed workload will appear,
it will be possible that the saturation length is still too big
for mixed workload. This would mean that the disk is too slow
(or the latency goal is too small), but for now this strategy
makes things better than they were.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>