io_queue: Move pieces of priority_class_data inline
The account_for method will soon be surrounded with its peers
that will do accounting on various request lifetime events.
It will be less code if keeping these beads compact.

The constructor is quite small by itself and will be patched
too, so to reduce the size of next patch move it inline too.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>