seastar-addr2line: support oneline backtrace in resolve call
Move splitting of the processed line to BacktraceResolver.__call__
so that it can match all the addressed on a single line
(that might be prefixed by non-address text).

Define BacktraceParser to categorize each line
as either None, SEPARATOR, or ADDRESS and return
the parsed results in a dictionary containing
the type, and for ADDRESS: an optional prefix and one or more addresses,
each represented by a dictionary containing optional path, and
an address.

If a line contains multiple addresses, it is considred as a complete
backtrace, hence it first prints the current backtrace (if any)
and prints one immediately after processing the line.

In addition, a self-test was added, that runs the BacktraceParser
and a bunch of possible lines and verifies the expected result.

Fix #892

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>