io_tester: Grab metrics and report back
When workload finishes the tester reports what was explicitly
measured -- number of operations, bytes processed and total observed
latency.

This patch makes it also report the time requests spent in queue and
in disk as well as the number of requests pushed through it. The
latter might differ slightly from the test's numbers, because test
ignores the requests that fell out of the testing timeframe, but
those outliers would have been accounted in the metrics.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>