tests: IO scheduler manual test
The new test checks if the IO sched does its job -- makes requests
not spend more time in disk than it's configured. To get this time
the recently introduced in-disk metrics is used.

The test exploits so called "saturation" test -- it starts with the
"base" minimally possible workload to make sure that the disk is able
to provide the desired latency at all. Then the workload is scaled up
(increasing the concurrency) and the test checks if the in-disk time
remains within bounds. Scaling stops once the iosched is saturated
which is -- starts to hold requests on in the queue. The latter is
estimated by the delay metrics. For a bit of extra confidence the
saturation is chedked for 2 more larger workloads.

Currently the test probes the scheduler with 5 types of workloads:

- pure 128k reads
- pure 4k writes
- pure 64k writes
- mixed 128k reads vs 4k writes
- mixed 128k reads vs 64k writes

The mixed workloads are scaled in 3 directions -- 1 symmetrical, when
both increase their concurrency; and 2 "alog the axis", when only one
part scales up and the other stays one-at-a-time.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>