reactor: disable xfs extent size hints if using the kernel page cache
The XFS code has this comment:

    /* we can't use delayed allocations when using extent size hints */

Delayed allocations are a great optimization when using the kernel
page cache, since it a delayed allocation can be elided if the file
is removed shortly after being written, which is the case for test
suites (the primary target of --kernel-page-cache). I observed tons
of disk flushes due to this; they go away if I disable extent size
hints.

Note that extent size hints are also unnecessary in the presence of
delayed allocation; the kernel will know the size of the extent to
allocate after the writes are ready to be sent to disk.

Fix the extra fsyncs by disabling the extent size hint if the
kernel page cache is enabled.
Message-Id: <20210510135717.2800708-1-avi@scylladb.com>