weak_ptr: lift restriction on copying
Add a copy constructor and copy assignment operator. No real drawback
for having them, but their lack can severely impact usability.
Copying a weak_ptr is conceptually equivalent to:

    auto copy = my_weak_ptr ? my_weak_ptr->weak_from_this() : nullptr;

A side-effect of this patch is that the `new (this) weak_ptr` idiom
which has disputed legality is gotten rid of.

Signed-off-by: Botond Dénes <bdenes@scylladb.com>
Message-Id: <20210513111602.225890-1-bdenes@scylladb.com>