memory: initialize cpu_mem access pointer eagerly, not lazily
cpu_mem is an expensive variable to use, since it is thread-local
and has non-trivial constructors and destructors. To work around
the compiler thread-local guard variable checks, we have get_cpu_mem
which is supposed to be inlined and the initialization optimized,
but it was observed not to be when we increased complexity.

To fix this, eagerly initialize the cpu_mem pointer, making
get_cpu_mem() trivial. It can now be inlined without pulling
in cpu_mem initialization.

This was made possible by the new checks for is_reactor_thread
that give us an opportunity for eager initialiation (calling a
new function init_cpu_mem()).