memory: fast-path local statistics
We keep two different kinds of statistics: thread-local, for reactor
threads, and global/hashed, for alien threads. The increment() function
decides what type it needs to use.

Reduce instruction count by splitting the local option into a new
increment_local() function, and call it directly when it is known
that it is the correct choice.