alien: prepare for multi-instance use
As a continuation for e6463df8a ("smp: allow having multiple instances
of the smp class"), alien is extended to support multiple smp instances.
This means several independent Seastar shard-sets running in the same
process, to allow single-process testing of clusters.

To do this, the alien::smp class which was just a container of some
static functions and variables is renamed to alien::instance, and
its functions and variables are made into non-static members.

The alien::run_on() and alien::submit_to() functions are modified to
accept an alien::instance parameter, so they know which Seastar instance
to target (the old signatures are retained but deprecated; they will target
the first Seastar intance created).

To obtain the alien::instance object, one can call app_template::alien().
To construct multiple Seastar instances, one can create several app_template
instances and run them.

Note: qs_deleter was made into a non-member class. As a private member
it was not std::is_default_constructible, which made the unique_ptr
constructor using it not accessible.
Message-Id: <20210603173538.3843447-1-avi@scylladb.com>