reactor: fix race with boost::barrier destructor during smp initialialization
In e6463df8a03 ("smp: allow having multiple instances of the smp class"),
we made some boost::barrier objects local to smp::configure. This leads
to the following race:

 1. The last thread to reach the barrier signals that it did so. This
    involves locking, signalling, and then unlocking. The unlocking has
    not been done yet.
 2. The main thread, waiting on the same barrier, continues execution
 3. The main thread reaches the end of smp::configure() and destroys
    those barriers
 4. The main thread crashes since the barrier still has a lock in use
 5. The last thread to reach the barrier unlocks the barrier. This doesn't
    happen since we crashed in step 4.

The fix is to use std::shared_ptr with the barrier, so that the thread to
complete work will destroy the barrier, rather than the main thread.
Message-Id: <20210607092207.264544-1-avi@scylladb.com>