net/tls: on TLS handshake failure, send error to client
If we configure TLS to accept only TLSv1.2, then a client trying to connect
with an earlier version of TLS will, rightly, fail to connect. However,
currently the client will just see the connection abruptly disconnected,
without getting any error message. For example, Python's SSL functions
report that "EOF occurred in violation of protocol" - which tells the
client nothing about the nature of the problem.

The problem is that the gnutls_handshake() function which we use in our
TLS implementation doesn't return a reply to the client. We need to use
gnutls_alert_send() to send such a reply. This is what this patch does.

We allow the gnutls function gnutls_error_to_alert() to pick the alert
to send for a given error in gnutls_handshake(). It seems that this
function picks reasonable alerts: In the unsupported protocol version
example, it picks the "protocol_version" alert, like the RFC suggests.
Python's SSL later prints this alert as "[SSL: TLSV1_ALERT_PROTOCOL_VERSION]
tlsv1 alert protocol version". This is perhaps not the clearest error
message in the world, but much better and easier to understand that the
EOF message we had before this patch.

Refs scylladb/scylla#8827
Fixes #912

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <20210609173357.1663154-1-nyh@scylladb.com>