reactor: Fix calculations of bandwidth in legacy mode
When the --max-io-requests option is provided the io-scheduler tries
to configure the fair-queue limits from it. To do it it takes into
account that fair-queue tickets are biased by multiplication on the
read_request_base_count parameter (to differentiate reads vs writes).
The problem is that when calculating the bandwidth the formula uses
the multiplied value of the maximum number of requests and multiplies
it again. This is wrong and leads to overflows, which in turn can
put 0 into the bandwidth locking the queue.

The fix is to calculate the bandwidth from the unbiased max number of
reqests and a static assertion that this won't overflow the unsigned
type either.

fixes: #917

v2:
- use common variable to get max-req-count from capacity
  and aio-max-nr

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>
Message-Id: <20210705184356.28839-1-xemul@scylladb.com>