testing: Lazily create seastar::app thread
When boost testing framework is initialized it calls
test_runner::start() which creates a thread for seastar app and
starts this app inside. The app creation in seastar is deadly
slow (as compared to plain execv()), the main time hogs are
cgroups reading code and hwloc topology initialization. The
rest is faster, but altogether consumes time as well.

This is not a problem per se, but sometimes the binary is not
started to execute a test, but to get some info about itself,
e.g. the --list_content option that just lists the cases. In
that case starting the thread with app is pointless, as it
will be shut down instantly. The shutdown process is notably
faster, but anyway takes time.

This patch delays the thread and app initialization until the
first test case is started. Since test cases are not run in
parallel inside one binary, the "first test case" can be safely
detected with the plain check.

Quick --list_content is very important for parallel test cases
running patchsed -- in dev mode listing cases for all 180+ tests
is ~20 seconds. With this patch it becomes 1.5.

tests: unit(dev), scylla.unit(dev)

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>
Message-Id: <20210709173622.31384-1-xemul@scylladb.com>