reactor: destroy_scheduling_group: make sure scheduling_group is valid
When called with an invalid sg (e.g. a previously destroyed one)
with_scheduling_group segfaults when dereferencing _task_queues[sg._id]
which is nullified by destroy_scheduling_group.

To prevent that, verify first the the sg is valid, and
abort if it isn't.

While at it, also mark the method noexcept.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>