reactor: Make sure run() doesn't throw
The main reactor loop cannot suddenly abort its execution, because
otherwise any tasks sitting in it lose the ability to clean after
themselves and any further execution (of main()) is not safe.

As the first step mark run() noexcept and report any exception
thrown from inside the rest of the reactor. The report would look
like

  ERROR 2021-07-15 20:59:24,266 [shard 0] seastar - io_submit: Inappropriate ioctl for device
  exception running reactor main loop on shard 0.
  Backtrace:
    0x22a3848
    0x22b4842
    0x226b7a3
    0x226ad06
    0xc49ffc
    /lib64/libc.so.6+0x281e1
    0xc48eed

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>