reactor: reactor_backend_aio: allow tuning number of network iocbs
The number of network iocbs was fixed since the introduction of
the linux-aio backend at 10,000. This limits the number of inactive
connections to 10,000 (active connections aren't polled). Anything
above that will crash.

This patch allows tuning that number. We can't just change it,
because /proc/sys/fs/aio-max-nr also has to be adjusted and this
is out of our control. So we pass the responsibility on:

 - the user can set it with --max-networking-io-control-blocks
 - the application (which can arrange for aio-max-nr to be tuned)
   can change the default for that option

I verified (with strace) that both application override and user
override work (and that user overrides the application default)

Fixes #858

Message-Id: <20210715185426.1166134-1-avi@scylladb.com>