compat: implement coroutine_handle<void> for <experimental/coroutine> header
We carry a compatibility header for clang's benefit, but it is missing
a specialization for std::experimental::coroutine_handle<void>.

This patch adds the specialization (which lacks promise()/from_promise())
members) and adds the conversion operator to decay to coroutine_handle<void>.
This follows the standard definition.

This is needed for coroutine integration where we lack access to the
promise type, and prefer not to implement a templated await_suspend().
Message-Id: <20210719153842.419824-1-avi@scylladb.com>