fair_queue: Shuffle dispatching loop
Right now its a loop in a loop that uses statistics to check if
it needs to continue. After the change it's a single loop and
statistics is _just_ statistics that doesn't take part in loop
decisions.

This change also helps teaching the fair_queue's priority queue
use priority_class_data references instead of shared_ptrs.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>