io_queue: Use on_internal_error in io_queue
When a non-read/write request is queued into io-queue the latter
throws an error because it's not yet ready to process anything but
reads or writes. It's better to use on_internal_error() here because
such errors are not about to be "gracefully recovered", but rather
need a backtrace to debug.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>
Message-Id: <20210727054952.12026-1-xemul@scylladb.com>