resource: define hwloc_topology_holder
Initializing, loading and eventually destroying
the hwloc topology is repeated in the code.

Refactor these calls out of allocate() and
nr_processing_units() into class hwloc_topology_holder
that initalizes the topology and loads it via get().
It destroys the hwloc_topology, if needed, in the destructor.

Note that the destructor path corresponds to
the current deferred actions that are no except too.

Extract hwloc_topology_holder into resource.hh
so that it can be used in the following patch in
smp::configure.

Test: unit(dev)

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>