resource: keep a singlton hwloc_topology_holder in configuration
Pavel Emelyanov <xemul@scylladb.com> said:
> The call to make hwloc_topology is very expensive, as it scans lots of files in /proc and cgroups.
> Currently it's called 2 times during smp::configure() and is taking up to 30% of configuration time.
> It's perfectly worth to make this a singleton and load the topology once per smp::configure().

This change keeps a singlton instance of hwloc_topology_holder
in the resource::configure created in smp::configure and
pass a reference to the configuration both to allocate()
and now to nr_processing_units().

The first one called will load the hwloc_topology and the
other will just reuse it.

The hwloc_topology_t is destroyed along with
the resource::configuration instance when smp::configure exists.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>