cpu_stall_detector: add perf_event stall detector
The Linux perf_event subsystem can collect kernel traces, for when
the stall happens in the kernel, so add a new stall detector and
use it preferentially when available (it might not be available due
to settings of perf_event_paranoid.

The ioctl that resets the count does not appear to work 100%
reliably, so we detect false positives and ignore them.

Note we don't yet report kernel traces; this will come in the next
patch.