stall_detector: record and print kernel stack trace
Enable kernel stack collection. Note this can only be gathered via
the mmap interface, so mmap the ring buffer and use it.

I tested this against a large mmap(MAP_POPULATE) to force a kernel
stall, and it correctly reported the kernel trace. Note that it will
only report time when the thread is active (spin loops or real work).
If the task is preempted or asleep (say due to contention on a mutex)
this will not be reported here.