coroutine: add maybe_yield
Introduce coroutine::maybe_yield(), which can be used to indicate
a yield point in a coroutine that performs long computations without
other co_await calls. It is similar to `co_await make_ready_future<>()`
but is more descriptive.

The implementation is slightly simpler since we don't need to deal
with exceptions stored in the future, although I don't expect any
real benefits.
Message-Id: <20210905153730.3327485-1-avi@scylladb.com>