log: migrate to options api
Use statically declared structs for configuration instead of
boost::program_options.
In the process the `--help-logger` option is removed from the logger
options and it is instead fully handled by `app_template` itself.
The previous CLI API (`get_options_description()` and
`extract_settings()`) is left intact but is now backed by the new
`options` struct. This means that a few options will now use a different
type in the boost options descriptions. This can break applications that
meddled with the internal representation before (which they shouldn't
have done in the first place).