build: clarify choice of C++ dialect, and change default to C++20
Although Seastar officially supports only C++20 and C++17, the dialect
testing code in configure.py was outdated, checking for C++17 and C++14.
This on one hand misses a check for C++20, and on the other hand, may
accept a C++14-only compiler which will surely fail during the build.

So in this patch we fix the configure.py code to to check for C++20 and
C++17, in this order.

One consequence of this change is that a "configure.py" without an
explicit "--c++-dialect" option will now default to C++20, whereas
before it defaulted to C++17. I think this change is desirable, and
acceptable: It is desirable because C++17 support may go away in a
couple of years so new users should prefer C++20, not C++17. And it is
acceptable because users who rely on a specific C++ dialect should
choose it explicitly in their build, and not rely on the default choice.
This default choice kept changing over the years anyway (earlier it
was C++14, today it is C++17, and after this patch it is C++20).

Another change in this patch is that if none of the supported dialects
work, configure.py exits with an error message, suggesting to upgrade
the compiler or choose a C++ dialect explicitly. The previous behavior
of configure.py was was that the last dialect listed was used despite
not working - and the user would see an error from the compiler, which
was less helpful in understanding how to fix the problem.

Finally, this patch updates README.md on what is the default dialect
chosen by configure.py - this information was ambigous (or even misleading)
before this patch.

Fixes #943.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <20210912092848.67624-1-nyh@scylladb.com>